// src/app/claim-processing/manual-claim-processing/process-claim-no-pay-eob/model/nopay-bill-lines.model.ts
/**
 * Model class NopayBillLines
 * Path: screenbean/procclmnopayeob
 * Model: com::uhc::aarp::fox::domain::screenbean::procclmnopayeob::NopayBillLines
 * Legacy Mapping: NOPAY-BILL-LINES
 */
export class NopayBillLines {
  screenPlanCode = '';
  screenTosCode = '';
  screenProvider = '';
  screenFdos = '';
  screenTdos = '';
  screenAmt = '';
  screenAssign = '';
  validateFdos = '';
  validateTdos = '';
  serialNo?: number;
}