// src/app/claim-processing/manual-claim-processing/process-claim-no-pay-eob/model/proc-clm-no-pay-eob.model.ts
import {NopayBillLines} from './nopay-bill-lines.model';
/**
 * Model class ProcClmNoPayEob
 * Path: screenbean/procclmnopayeob
 * Model: com::uhc::aarp::fox::domain::screenbean::procclmnopayeob::ProcClmNoPayEob
 */
export class ProcClmNoPayEob {
  nopayBillLines: NopayBillLines[] = [];
  screenErr = '';
  wholeName = '';
  accountNo = '';
  ma2cmnd = '';
  ma2mem1 = '';
  ma2mem2 = '';
  ma2mem3 = '';
  ma2cno1 = '';
  ma2cno2 = '';
  ma2cno3 = '';
  ma2cno4 = '';
  ma2cno5 = '';
  ma2cno6 = '';
  ma2com1 = '';
  ma2com2 = '';
}